backend/requirements.txt
# FastAPI 核心
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# 数据库
psycopg2-binary==2.9.9
neo4j==5.14.1
pymilvus==2.3.4

# 语音处理
openai-whisper==20231117
vosk==0.3.45
azure-cognitiveservices-speech==1.32.1
edge-tts==6.1.9

# AI 和向量
openai==1.3.7
langchain==0.0.350
langchain-community==0.0.10
sentence-transformers==2.2.2

# 中文 NLP（用于实体识别）
jieba==0.42.1
pkuseg==0.0.25

# 工具库
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
httpx==0.25.2
aiofiles==23.2.1

# 数据处理
numpy==1.24.3
pandas==2.1.3

